restructed, material ui button
