+++ client/src/index.tsx
@@ -1,6 +1,7 @@
 import { StrictMode } from 'react';
 import { render } from 'react-dom';
-import App from './app';
+import './styles/index.css';
+import App from './App';
 
 render(
   <StrictMode>
